Переименовал папку с тестовым проектом
